Facebook Custom Audience Pixel Code
How it Works

The Custom Audience pixel is activated every time someone opens a webpage where the code is installed. When the pixel is activated an event is sent to Facebook's servers with general information about the browsing session including page URLs. The events are then compared to a set of Audience rules the advertiser has created. If the event matches an Audience rule the person who visited the webpage is added to that Audience. These website Custom Audiences can later be targeted by ads.

https://developers.facebook.com/docs/ads-for-websites/website-custom-audiences/getting-started#how-it-works